refactor: extract sandbox command shape tests
